Make the following changes:
 - Separate histogram fill tests from utilities

 - Add helper impl tuple limit
