+++ test/core/histogram/Jamfile
@@ -12,6 +12,7 @@ compile constructor.cpp ;
 compile dimension.cpp ;
 run access.cpp ;
 run cumulative.cpp ;
+run fill.cpp ;
 run hash_tuple.cpp ;
 run helpers.cpp ;
 run is_compatible.cpp ;
